requirements.txt
requests==2.32.3
feedparser==6.0.12
PyYAML==6.0.2